fix other tooltip errors
solve error of the tooltip boolean error, when pressing button submit
twice the tooltip hid.
delete the animation when using tooltip boolean since this was causing
the previous bug.
fix the input error, when the input have errors in the signup the bubble
error and the normal error were shown at the same time